Refactor UserPrCard component and add UserPrCard to CompetitionDetails page. Add some label based on github
